src/web/requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework
Flask==2.1.2
Flask-WTF==1.0.1

requests==2.28.0
python-dateutil==2.8.2
python-jose[cryptography]==3.3.0
